fix: initialise modules after scene load
