+++ public/index.html
@@ -206,7 +206,7 @@
   </a-scene>
 
   <!-- Client logic is loaded last once the DOM and libraries are ready -->
-  <script src="js/mingle_client.js"></script>
+  <script type="module" src="js/mingle_client.js"></script>
   <script src="js/mingle_navbar.js"></script>
 </body>
 </html>
